+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.43.0]
 - Remove ILIAS 5.4 support
+- Remove deprecated MultiLineInputGUI
 
 ## [0.42.9]
 - Hide HiddenInputGUI in MultiLineNewInputGUI